making blueprints
