+++ project/users/templates/users/user_favorites.html
